docs: address initial feedback
