+++ docs/content/index.md
@@ -30,8 +30,8 @@ fn hello_world(req: Request) -> Result<Response> {​
 Spin applications are comprised of one or more function-based _components_, and
 follow the event-driven model — they are executed as the result of events being
 generated by _triggers_ (for example an HTTP server receiving requests, or a queue
-subscription receiving messages). On each new event, _the entrypoint_ of a
-component is executed by Spin. The entrypoints to components are _functions_.
+subscription receiving messages). On each new event, _the entry point_ of a
+component is executed by Spin. The entry points to components are _functions_.
 This, together with the fact that they are invoked in response to events, brings
 the Spin application model closer to the Function-as-a-Service model.
