+++ docs/content/rust-components.md
@@ -14,6 +14,16 @@ official resources for learning Rust](https://www.rust-lang.org/learn).
 
 > All examples from this page can be found in [the Spin repository on GitHub](https://github.com/fermyon/spin/tree/main/examples).
 
+Besides `cargo` and the Rust compiler, you need to add the `wasm32-wasi` target
+before compiling Rust components for Spin:
+
+In order to compile Rust programs to Spin components, you also need the
+`wasm32-wasi` target. You can add it using `rustup`:
+
+```plaintext
+$ rustup target add wasm32-wasi
+```
+
 ## HTTP components
 
 In Spin, HTTP components are triggered by the occurrence of an HTTP request, and
@@ -45,7 +55,7 @@ fn hello_world(req: Request) -> anyhow::Result<Response> {
 The important things to note in the implementation above:
 
 - the `spin_sdk::http_component` macro marks the function as the
-  entrypoint for the Spin component
+  entry point for the Spin component
 - the function signature — `fn hello_world(req: Request) -> Result<Response>` —
   the Spin HTTP component uses the HTTP objects from the popular Rust crate
   [`http`](https://crates.io/crates/http), and the request and response bodies
@@ -144,7 +154,7 @@ fn on_message(msg: Bytes) -> Result<()> {
 ```
 
 - the `spin_sdk::redis_component` macro marks the function as the
-  entrypoint for the Spin component
+  entry point for the Spin component
 - in the function signature — `fn on_message(msg: Bytes) -> anyhow::Result<()>` —
 `msg` contains the payload from the Redis channel
 - the component returns a Rust `anyhow::Result`, so if there is an error
@@ -212,3 +222,21 @@ be used when implementing the component.
 
 > Make sure to read [the page describing the HTTP trigger](/http-trigger) for more
 > details about building HTTP applications.
+
+## Troubleshooting
+
+Sometimes things can go wrong, especially such early projects. If you bump into
+issues building and running your Rust component:
+
+- ensure `cargo` is present in your path — we recommend
+[Rust](https://www.rust-lang.org/) at [1.56+](https://www.rust-lang.org/tools/install)
+- ensure `wasm32-wasi` target is configured for your Rust installation —
+you can add it by running `rustup target add wasm32-wasi`
+- build a `release` version of the component — all Spin application definitions
+(`spin.toml` files) reference build configuration for Rust, so make sure to
+run `cargo build --release --target wasm32-wasi` when building your components
+(you can validate modules are correctly being built by checking the contents of
+the `target/wasm32-wasi/release` directory and looking for `.wasm` files)
+- make sure the path and name of the Wasm module in `target/wasm32-wasi/release`
+match `source` field in the component configuration (the `source` field contains
+the path to the Wasm module, relative to `spin.toml`)
