+++ docs/static/js/main.js
@@ -27,7 +27,7 @@ const systemTheme = window.matchMedia("(prefers-color-scheme: dark)").matches
   ? "dark"
   : "light";
 
-// the default theme is the system theme, unless the user has
+// The default theme is the system theme, unless the user has
 // explicitly overriden it.
 var savedTheme = localStorage.getItem("theme") || systemTheme;
 setTheme(savedTheme);
@@ -41,7 +41,7 @@ btn.addEventListener("click", () => {
   }
 });
 
-// change the website theme when the system theme changes.
+// Change the website theme when the system theme changes.
 window
   .matchMedia("(prefers-color-scheme: dark)")
   .addEventListener("change", (event) => {
@@ -51,6 +51,13 @@ window
       setTheme("light");
     }
   });
+
+// Change the website theme when a different tab changed it.
+window.onstorage = () => {
+  // When local storage changes, update the theme.
+  setTheme(localStorage.getItem("theme"));
+};
+
 
 function setTheme(mode) {
   localStorage.setItem("theme", mode);
@@ -63,7 +70,7 @@ function setTheme(mode) {
 }
 
 document.addEventListener("DOMContentLoaded", function(){
-  // init after dom
+  // Initialize after the DOM.
   (function () {
     var burger = document.querySelector('.burger');
     var menu = document.querySelector('#' + burger.dataset.target);
