+++ docs/templates/content_sidebar.hbs
@@ -4,7 +4,7 @@
             Using Spin
         </p>
         <ul class="menu-list">
-            <li><a href="{{site.info.base_url}}/index/">Introduction</a></li>
+            <li><a href="{{site.info.base_url}}/">Introduction</a></li>
             <li><a href="{{site.info.base_url}}/quickstart/">Quickstart</a></li>
             <li><a href="{{site.info.base_url}}/configuration/">Configuration</a></li>
         </ul>
@@ -15,6 +15,7 @@
         <ul class="menu-list">
             <li><a href="{{site.info.base_url}}/rust-components/">Rust</a></li>
             <li><a href="{{site.info.base_url}}/go-components/">Go</a></li>
+            <li><a href="{{site.info.base_url}}/other-languages/">Other languages</a></li>
         </ul>
 
         <p class="menu-label">
