--- examples/spin-timer-echo/echo.wit
@@ -1,1 +1,0 @@
-echo: function(msg: string) -> string
