--- examples/spin-timer-echo/example/src/lib.rs
@@ -1,8 +1,0 @@
-wit_bindgen_rust::export!("../echo.wit");
-
-struct Echo;
-impl echo::Echo for Echo {
-    fn echo(msg: String) -> String {
-        format!("ECHO: {}", msg)
-    }
-}
