+++ examples/spin-timer/spin-timer.wit
@@ -1,0 +1,1 @@
+handle-timer-request: function(msg: string) -> string
